refactor(cli): split up loadWorldConfig into multiple files
