+++ packages/cli/src/commands/deploy-v2.ts
@@ -2,7 +2,7 @@ import chalk from "chalk";
 import glob from "glob";
 import path, { basename } from "path";
 import type { CommandModule } from "yargs";
-import { loadWorldConfig } from "../config/loadWorldConfig.js";
+import { loadWorldConfig } from "../config/world/index.js";
 import { deploy } from "../utils/deploy-v2.js";
 import { logError, MUDError } from "../utils/errors.js";
 import { forge, getRpcUrl, getSrcDirectory } from "../utils/foundry.js";
