+++ packages/cli/src/config/index.ts
@@ -1,5 +1,5 @@
 import { StoreUserConfig, StoreConfig } from "./parseStoreConfig.js";
-import { WorldUserConfig, ResolvedWorldConfig } from "./loadWorldConfig.js";
+import { WorldUserConfig, ResolvedWorldConfig } from "./world/index.js";
 
 export type MUDUserConfig = StoreUserConfig & WorldUserConfig;
 export type MUDConfig = StoreConfig & ResolvedWorldConfig;
@@ -8,5 +8,5 @@ export * from "./commonSchemas.js";
 export * from "./loadConfig.js";
 export * from "./loadStoreConfig.js";
 export * from "./parseStoreConfig.js";
-export * from "./loadWorldConfig.js";
+export * from "./world/loadWorldConfig.js";
 export * from "./validation.js";
