+++ packages/cli/src/index.ts
@@ -1,6 +1,6 @@
 export { parseStoreConfig } from "./config/parseStoreConfig.js";
 export { loadStoreConfig } from "./config/loadStoreConfig.js";
-export { loadWorldConfig, resolveWorldConfig, parseWorldConfig } from "./config/loadWorldConfig.js";
+export { loadWorldConfig, resolveWorldConfig, parseWorldConfig } from "./config/world/loadWorldConfig.js";
 export { renderTablesFromConfig } from "./render-solidity/renderTablesFromConfig.js";
 export { renderTable } from "./render-solidity/renderTable.js";
